Members directory using handsontable
