+++ resources/views/layouts/app.blade.php
@@ -66,11 +66,15 @@
                 <ul class="nav navbar-nav">
                     <li><a href="{{ url('/home') }}">Home</a></li>
                     @if (!is_null(Auth::user()) and Auth::user()->hasRole('viewer'))
+
                     <li><a href="{{ route('comm.members') }}">Members DB</a></li>
                     @endif
+
                     @if (!is_null(Auth::user()) and Auth::user()->hasRole('editor'))
                     <li><a href="{{ url('/admin') }}">Edit Members</a></li>
                     @endif
+
+                    <li><a href="{{ route('member.directory') }}">Members Directory</a></li>
                 </ul>
 
                 <!-- Right Side Of Navbar -->
